requirements.txt
pandas
numpy
nltk
scikit-learn==0.24.2
collection
matplotlib
dvc
dvc[gdrive]
pyyaml
